+++ gamemodes/terrortown/gamemode/client/cl_main.lua
@@ -316,8 +316,6 @@ end
 -- @ref https://wiki.facepunch.com/gmod/GM:PostGamemodeLoaded
 -- @local
 function GM:PostGamemodeLoaded()
-	events.OnLoaded()
-
 	ScoringEventSetup()
 end
 
@@ -332,8 +330,6 @@ function GM:OnReloaded()
 
 	local skinName = vskin.GetVSkinName()
 	vskin.UpdatedVSkin(skinName, skinName)
-
-	events.OnLoaded()
 
 	ScoringEventSetup()
 end