+++ gamemodes/terrortown/gamemode/shared/sh_init.lua
@@ -503,28 +503,31 @@ require("outline")
 include("includes/modules/pon.lua")
 
 -- include extensions
+include("ttt2/extensions/math.lua")
 include("ttt2/extensions/net.lua")
 include("ttt2/extensions/sql.lua")
 include("ttt2/extensions/string.lua")
 include("ttt2/extensions/table.lua")
+include("ttt2/extensions/util.lua")
 include("ttt2/extensions/surface.lua")
 include("ttt2/extensions/draw.lua")
 
 -- include libraries
+include("ttt2/libraries/fileloader.lua")
+include("ttt2/libraries/classbuilder.lua")
 include("ttt2/libraries/fonts.lua")
 include("ttt2/libraries/appearance.lua")
 include("ttt2/libraries/drawsc.lua")
 include("ttt2/libraries/vguihandler.lua")
 include("ttt2/libraries/vskin.lua")
-include("ttt2/libraries/fileloader.lua")
 include("ttt2/libraries/door.lua")
 include("ttt2/libraries/orm.lua")
 include("ttt2/libraries/thermalvision.lua")
 include("ttt2/libraries/events.lua")
-
+include("ttt2/libraries/none.lua")
+include("ttt2/libraries/targetid.lua")
 
 -- include ttt required files
-ttt_include("sh_util")
 ttt_include("sh_decal")
 ttt_include("sh_lang")
 ttt_include("sh_sql")