+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
   - added a hook `TTT2AddedEvent` that is called after an event was added
 - Added `orm` library to simplify database access
 - Added French translation (by @MisterClems)
+- Added a new classbuilder that can be used to create classes from files
 
 ### Changed
 
@@ -34,11 +35,13 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 - Changed `TTTCanPickupAmmo` hook, return `false` to cancel pickup
 - Changed `TTTPlayerUsedHealthStation` hook, return `false` to cancel health regeneration tick
 - Changed all C4 hooks to be cancelable
+- moved functions from sh_util into their respective library files
 
 ### Fixed
 
 - Fixed weapon pickup bug, where weapons would not get dropped but stayed in inventory
 - Fixed defuser only working for detectives
+- Fixed a roleselection bug, where forced roles would not be deducted from the available roles
 
 ## [v0.7.4b](https://github.com/TTT-2/TTT2/tree/v0.7.4b) (2020-09-28)
 
@@ -49,6 +52,7 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 - Added `table.ExtractRandomEntry(tbl, filterFn)` function
 - Added a team indicator in front of every name in the scoreboard (just known teams will be displayed)
 - Added a hook `TTT2ModifyCorpseCallRadarRecipients` that is called once "call detective" is pressed
+- Added a `targetid` library, that can be used to draw TargetIDs for entities
 
 ### Changed